# app/routers/agents.py
"""Agent routes — context, learning, knowledge graph."""

import uuid
from typing import Optional
from fastapi import APIRouter, Request, Depends, Query
from sqlalchemy import text
from sqlalchemy.ext.asyncio import AsyncSession

from app.db import get_db
from pydantic import BaseModel, Field
from app.models.api import AgentLearnRequest, MemorySaveRequest, MemorySaveResponse, AgentContextResponse, AgentTransferResponse


class EntityItem(BaseModel):
    name: str
    entityType: str = ""

class EntitiesRequest(BaseModel):
    entities: list[EntityItem] = Field(default_factory=list)

class RelationItem(BaseModel):
    from_: str = Field(alias="from")
    to: str
    relationType: str = ""
    model_config = {"populate_by_name": True}

class RelationsRequest(BaseModel):
    relations: list[RelationItem] = Field(default_factory=list)
from app.agents.peer import get_or_create
from app.agents.context import get_context
from app.memory.ingest import save_memory

router = APIRouter()


@router.get("/{agent_id}/context", response_model=AgentContextResponse)
async def agent_context(
    agent_id: str,
    request: Request,
    db: AsyncSession = Depends(get_db),
    tokens: int = Query(default=2000, ge=100, le=16000),
    search_query: Optional[str] = Query(default=None),
):
    await get_or_create(db, agent_id)
    return await get_context(db, agent_id, token_budget=tokens, search_query=search_query)


@router.post("/{agent_id}/learn", response_model=MemorySaveResponse)
async def agent_learn(
    agent_id: str,
    body: AgentLearnRequest,
    request: Request,
    db: AsyncSession = Depends(get_db),
):
    req = MemorySaveRequest(
        content=body.content,
        agent_id=agent_id,
        memory_type=body.memory_type,
        importance=body.importance,
        metadata=body.metadata,
        tags=body.tags,
    )
    return await save_memory(db, request.app.state.redis, req)


@router.post("/{from_agent}/transfer/{to_agent}", response_model=AgentTransferResponse)
async def transfer_agent_knowledge(
    from_agent: str,
    to_agent: str,
    request: Request,
    db: AsyncSession = Depends(get_db),
    limit: int = Query(default=10, ge=1, le=50),
):
    """Copy recent high-value memories from one agent to another."""
    await get_or_create(db, from_agent)
    await get_or_create(db, to_agent)

    rows = await db.execute(text("""
        SELECT m.content, m.memory_type, m.importance, m.metadata
        FROM memories m
        JOIN agents a ON a.id = m.agent_id
        WHERE a.name = :from_agent
        ORDER BY m.importance DESC, m.created_at DESC
        LIMIT :limit
    """), {"from_agent": from_agent, "limit": limit})

    transferred_ids: list[str] = []
    for row in rows.fetchall():
        metadata = dict(row.metadata or {})
        tags = list(metadata.get("tags") or [])
        if "transferred" not in tags:
            tags.append("transferred")
        metadata.update({"transferred_from": from_agent, "tags": tags})
        req = MemorySaveRequest(
            content=row.content,
            agent_id=to_agent,
            memory_type=row.memory_type,
            importance=min(1.0, float(row.importance or 0.5) + 0.05),
            metadata=metadata,
            tags=tags,
        )
        saved = await save_memory(db, request.app.state.redis, req)
        transferred_ids.append(saved.id)

    return AgentTransferResponse(
        from_agent=from_agent,
        to_agent=to_agent,
        transferred=len(transferred_ids),
        memory_ids=transferred_ids,
    )


@router.post("/{agent_id}/entities")
async def upsert_entities(agent_id: str, body: EntitiesRequest, db: AsyncSession = Depends(get_db)):
    """Create knowledge graph entities (skips existing names)."""
    await get_or_create(db, agent_id)
    agent_row = await db.execute(text("SELECT id FROM agents WHERE name = :n"), {"n": agent_id})
    aid = agent_row.scalar()
    created = []
    for ent in body.entities:
        name = ent.name.strip()
        if not name:
            continue
        result = await db.execute(text("""
            INSERT INTO entities (name, entity_type, agent_id)
            VALUES (:name, :etype, :aid)
            ON CONFLICT DO NOTHING
            RETURNING id
        """), {"name": name, "etype": ent.entityType, "aid": aid})
        row = result.fetchone()
        if row:
            created.append({"name": name, "id": str(row.id)})
    await db.commit()
    return {"created": len(created), "entities": created}


@router.post("/{agent_id}/relations")
async def upsert_relations(agent_id: str, body: RelationsRequest, db: AsyncSession = Depends(get_db)):
    """Create knowledge graph relations (skips duplicates)."""
    await get_or_create(db, agent_id)
    agent_row = await db.execute(text("SELECT id FROM agents WHERE name = :n"), {"n": agent_id})
    aid = agent_row.scalar()
    created = 0
    for rel in body.relations:
        frm = rel.from_.strip()
        to = rel.to.strip()
        rtype = rel.relationType.strip()
        if not (frm and to and rtype):
            continue
        # Ensure both entities exist
        for ename in (frm, to):
            await db.execute(text("""
                INSERT INTO entities (name, entity_type, agent_id)
                VALUES (:name, 'unknown', :aid)
                ON CONFLICT DO NOTHING
            """), {"name": ename, "aid": aid})
        fe = await db.execute(text("SELECT id FROM entities WHERE name = :n AND agent_id = :aid"), {"n": frm, "aid": aid})
        te = await db.execute(text("SELECT id FROM entities WHERE name = :n AND agent_id = :aid"), {"n": to, "aid": aid})
        fid = fe.scalar()
        tid = te.scalar()
        if fid and tid:
            await db.execute(text("""
                INSERT INTO relations (from_entity_id, to_entity_id, relation_type)
                VALUES (:fid, :tid, :rtype)
                ON CONFLICT DO NOTHING
            """), {"fid": fid, "tid": tid, "rtype": rtype})
            created += 1
    await db.commit()
    return {"created": created}


@router.get("/{agent_id}/entities")
async def get_entities(agent_id: str, names: str = Query(default=""), db: AsyncSession = Depends(get_db)):
    """Retrieve entities by name list (comma-separated)."""
    name_list = [n.strip() for n in names.split(",") if n.strip()]
    if not name_list:
        rows = await db.execute(text("""
            SELECT e.name, e.entity_type,
                   array_agg(DISTINCT r.relation_type || '->' || te.name) FILTER (WHERE te.name IS NOT NULL) as relations
            FROM entities e
            JOIN agents a ON a.id = e.agent_id AND a.name = :agent_id
            LEFT JOIN relations r ON r.from_entity_id = e.id
            LEFT JOIN entities te ON te.id = r.to_entity_id
            GROUP BY e.name, e.entity_type
            LIMIT 100
        """), {"agent_id": agent_id})
    else:
        rows = await db.execute(text("""
            SELECT e.name, e.entity_type,
                   array_agg(DISTINCT r.relation_type || '->' || te.name) FILTER (WHERE te.name IS NOT NULL) as relations
            FROM entities e
            JOIN agents a ON a.id = e.agent_id AND a.name = :agent_id
            LEFT JOIN relations r ON r.from_entity_id = e.id
            LEFT JOIN entities te ON te.id = r.to_entity_id
            WHERE e.name = ANY(:names)
            GROUP BY e.name, e.entity_type
        """), {"agent_id": agent_id, "names": name_list})
    result = []
    for row in rows.fetchall():
        result.append({
            "name": row.name,
            "entityType": row.entity_type,
            "relations": [r for r in (row.relations or []) if r],
        })
    return {"entities": result}


@router.get("/{agent_id}/blast-radius")
async def blast_radius(
    agent_id: str,
    files: str = Query(default=""),
    max_depth: int = Query(default=2, ge=1, le=5),
    db: AsyncSession = Depends(get_db),
):
    """Find all entities impacted by changed files via BFS relation traversal."""
    file_list = [f.strip() for f in files.split(",") if f.strip()]
    if not file_list:
        return {"impacted": [], "depth": 0}
    # Seed: entities whose name matches any changed file (by path component)
    seeds = []
    for fname in file_list:
        basename = fname.split("/")[-1]
        rows = await db.execute(text("""
            SELECT e.id, e.name FROM entities e
            JOIN agents a ON a.id = e.agent_id AND a.name = :agent_id
            WHERE e.name ILIKE :pat OR e.name ILIKE :base
        """), {"agent_id": agent_id, "pat": f"%{fname}%", "base": f"%{basename}%"})
        seeds.extend(rows.fetchall())

    visited = {str(r.id): r.name for r in seeds}
    frontier = list(visited.keys())
    depth = 0
    while frontier and depth < max_depth:
        rows = await db.execute(text("""
            SELECT DISTINCT e.id::text, e.name, r.relation_type
            FROM relations r
            JOIN entities e ON e.id = r.to_entity_id
            WHERE r.from_entity_id = ANY(CAST(:frontier AS uuid[]))
            AND r.to_entity_id != ALL(CAST(:visited AS uuid[]))
        """), {"frontier": frontier, "visited": list(visited.keys())})
        new_frontier = []
        for row in rows.fetchall():
            if row.id not in visited:
                visited[row.id] = row.name
                new_frontier.append(row.id)
        frontier = new_frontier
        depth += 1

    return {
        "changed_files": file_list,
        "impacted": [{"id": k, "name": v} for k, v in visited.items()],
        "depth_reached": depth,
    }